requirements.txt
# torch Optional (uncomment if using PyTorch)

# Core FastAPI and Server
fastapi==0.110.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Database - MongoDB
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6

# Data Processing & Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# OpenRouter/LLM Integration (if used in ai_service.py)
# openai==1.6.1
httpx==0.28.1
aiohttp==3.9.1

# Utilities
requests==2.31.0

# Performance & Optimization
orjson==3.9.10

# Compatibility for async libraries
anyio==4.8.0

# Build dependencies for Python 3.12+
setuptools>=69.0.0
wheel>=0.42.0

pydantic[email]==2.5.2

joblib==1.3.2

pandas==2.1.4

google-generativeai==0.8.5
python-dotenv==1.0.1

bcrypt==4.0.1

scikit-learn==1.6.1

supabase==2.22.0

aiofiles==25.1.0

fairlearn==0.13.0

# Machine Learning
# statsmodel==0.14.5
# matplotlib==3.10.7
# seaborn==0.13.2
# shap==0.49.1